refactor: 마이그레이션 구현체 교체 및 소비처 정리
