--- packages/jds/src/components/Checkbox/migration/checkbox.types.ts
@@ -1,84 +1,0 @@
-import type { ComponentPropsWithoutRef, ReactNode } from "react";
-
-export const CHECKBOX_SIZE_OPTIONS = ["xs", "sm", "md", "lg"] as const;
-export const CHECKBOX_VARIANT_OPTIONS = ["hollow", "outlined"] as const;
-
-export type CheckboxSize = (typeof CHECKBOX_SIZE_OPTIONS)[number];
-export type CheckboxVariant = (typeof CHECKBOX_VARIANT_OPTIONS)[number];
-export type CheckedState = boolean | "indeterminate";
-
-// Checkbox.Root
-
-type CheckboxRootControlledProps = {
-  value: string[];
-  defaultValue?: never;
-  onChange: (value: string[]) => void;
-};
-
-type CheckboxRootUncontrolledProps = {
-  value?: never;
-  defaultValue?: string[];
-  onChange?: (value: string[]) => void;
-};
-
-type CheckboxRootBaseProps = {
-  size?: CheckboxSize;
-  variant?: CheckboxVariant;
-  disabled?: boolean;
-  isInvalid?: boolean;
-  name?: string;
-  children: ReactNode;
-};
-
-export type CheckboxRootProps = CheckboxRootBaseProps &
-  (CheckboxRootControlledProps | CheckboxRootUncontrolledProps);
-
-// Checkbox.Item
-
-export interface CheckboxItemProps extends ComponentPropsWithoutRef<"label"> {
-  size?: CheckboxSize;
-  variant?: CheckboxVariant;
-  disabled?: boolean;
-  isInvalid?: boolean;
-  children: ReactNode;
-}
-
-// Checkbox.Basic
-// Checkbox.Root 안에서 사용 시 value가 그룹 내 식별자가 되며 checked 상태는 그룹이 관리한다.
-
-type CheckboxBasicControlledProps = {
-  checked: CheckedState;
-  defaultChecked?: never;
-  onCheckedChange: (checked: CheckedState) => void;
-};
-
-type CheckboxBasicUncontrolledProps = {
-  checked?: never;
-  defaultChecked?: boolean;
-  onCheckedChange?: (checked: CheckedState) => void;
-};
-
-type CheckboxBasicBaseProps = Omit<
-  ComponentPropsWithoutRef<"input">,
-  "size" | "checked" | "defaultChecked" | "onChange"
-> & {
-  size?: CheckboxSize;
-  value?: string;
-  disabled?: boolean;
-  isInvalid?: boolean;
-};
-
-export type CheckboxBasicProps = CheckboxBasicBaseProps &
-  (CheckboxBasicControlledProps | CheckboxBasicUncontrolledProps);
-
-// Checkbox.Label
-
-export interface CheckboxLabelProps {
-  children: ReactNode;
-}
-
-// Checkbox.Helper
-
-export interface CheckboxHelperProps {
-  children: ReactNode;
-}
