+++ packages/jds/src/components/Dialog/Dialog.types.ts
@@ -2,13 +2,17 @@ import type * as DialogPrimitive from "@radix-ui/react-dialog";
 import type { ReactNode } from "react";
 
 import type { BlockButtonBasicProps } from "../Button/BlockButton";
-import type { CheckboxContentProps } from "../Checkbox";
+import type { CheckedState } from "../Checkbox";
 
 type BlockButtonActionProps = Pick<
   BlockButtonBasicProps,
   "children" | "onClick" | "disabled" | "hierarchy"
 >;
-type CheckBoxActionProps = Pick<CheckboxContentProps, "label" | "checked" | "onCheckedChange">;
+type CheckBoxActionProps = {
+  label: ReactNode;
+  checked: CheckedState;
+  onCheckedChange: (checked: CheckedState) => void;
+};
 type RadixDialogProps = Omit<DialogPrimitive.DialogProps, "children">;
 
 type DialogActionOptions =
